Moved EncodedSciData class to it's own file
Also removed old and unused headers from HammingBlock class
